finish the refactoring of Action and Observation, ready for gym integration now
