+++ docs/space.rst
@@ -1,0 +1,9 @@
+Space
+===================================
+.. automodule:: grid2op.Space
+    :members:
+    :show-inheritance:
+    :special-members:
+
+.. include:: final.rst
+
