+++ grid2op/Backend.py
@@ -160,305 +160,11 @@ class Backend(GridObjects, ABC):
         # if it's set to true, it returns all intermediate _grid states. This can slow down the computation!
         self.detailed_infos_for_cascading_failures = detailed_infos_for_cascading_failures
 
-        # self.n_line = None  # int: number of powerlines
-        # self.n_gen = None  # int: number of generators
-        # self.n_load = None  # int: number of loads
-        # self.n_sub = None  # int: number of substations
-        # self.sub_info = None  # vector[int]: of size number of substation. Tells for each substation the number of element connected to it
-        #
-        # self.load_to_subid = None  # vector[int]: as size number of load, giving for each the substation id to which it is connected
-        # self.gen_to_subid = None  # vector[int]: as size number of generators, giving for each the substation id to which it is connected
-        # self.line_or_to_subid = None  # vector[int]: as size number of lines, giving for each the substation id to which its "origin" end is connected
-        # self.line_ex_to_subid = None  # vector[int]: as size number of lines, giving for each the substation id to which its "extremity" end is connected
-        #
-        # # position in the vector of substation
-        # self.load_to_sub_pos = None   # vector[int]: as size number of load, giving for each the postition of this load in among the element of this substation
-        # self.gen_to_sub_pos = None
-        # self.line_or_to_sub_pos = None
-        # self.line_ex_to_sub_pos = None
-        #
-        # # position in the topological vector
-        # # for internal use only, set it with "_compute_pos_big_topo"
-        # # after having loaded the _grid.
-        # self.load_pos_topo_vect = None
-        # self.gen_pos_topo_vect = None
-        # self.line_or_pos_topo_vect = None
-        # self.line_ex_pos_topo_vect = None
-        # # see definition of "_compute_pos_big_topo" for more information about these vector
-        #
-        # # name of the objects
-        # self.name_load = None
-        # self.name_gen = None
-        # self.name_line = None
-        # self.name_sub = None
-
         # the power _grid manipulated. One powergrid per backend.
         self._grid = None
 
         # thermal limit setting, in ampere, at the same "side" of the powerline than self.get_line_overflow
         self.thermal_limit_a = None
-
-    def _aux_pos_big_topo(self, vect_to_subid, vect_to_sub_pos):
-        """
-        Return the proper "_pos_big_topo" vector given "to_subid" vector and "to_sub_pos" vectors.
-        This function is also called to performed sanity check after the load on the powergrid.
-
-        :param vect_to_subid: vector of int giving the id of the topology for this element
-        :type vect_to_subid: iterable int
-
-        :param vect_to_sub_pos: vector of int giving the id IN THE SUBSTATION for this element
-        :type vect_to_sub_pos: iterable int
-
-        :return:
-        """
-        res = np.zeros(shape=vect_to_subid.shape)
-        for i, (sub_id, my_pos) in enumerate(zip(vect_to_subid, vect_to_sub_pos)):
-            obj_before = np.sum(self.sub_info[:sub_id])
-            res[i] = obj_before + my_pos
-        return res
-
-    def _compute_pos_big_topo(self):
-        """
-        Compute the position of each element in the big topological vector.
-
-        Topology action are represented by numpy vector of size np.sum(self.sub_info).
-        The vector self.load_pos_topo_vect will give the index of each load in this big topology vector.
-        For examaple, for load i, self.load_pos_topo_vect[i] gives the position in such a topology vector that
-        affect this load.
-
-        This position can be automatically deduced from self.sub_info, self.load_to_subid and self.load_to_sub_pos.
-
-        This is the same for generators and both end of powerlines
-
-        :return: ``None``
-        """
-        # self.assert_grid_correct()
-        self.load_pos_topo_vect = self._aux_pos_big_topo(self.load_to_subid, self.load_to_sub_pos).astype(np.int)
-        self.gen_pos_topo_vect = self._aux_pos_big_topo(self.gen_to_subid, self.gen_to_sub_pos).astype(np.int)
-        self.line_or_pos_topo_vect = self._aux_pos_big_topo(self.line_or_to_subid, self.line_or_to_sub_pos).astype(np.int)
-        self.line_ex_pos_topo_vect = self._aux_pos_big_topo(self.line_ex_to_subid, self.line_ex_to_sub_pos).astype(np.int)
-
-    def assert_grid_correct(self):
-        """
-        Performs some checking on the loaded _grid to make sure it is consistent.
-        It also makes sure that the vector such as *sub_info*, *load_to_subid* or *gen_to_sub_pos* are of the
-        right type eg. numpy.array with dtype: np.int
-
-        It is called after the _grid has been loaded.
-
-        These function is by default called by the :class:`grid2op.Environment` class after the initialization of the environment.
-        If these tests are not successfull, no guarantee are given that the backend will return consistent computations.
-
-        In order for the backend to fully understand the structure of actions, it is strongly advised NOT to override this method.
-
-        :return: ``None``
-        :raise: :class:`grid2op.EnvError` and possibly all of its derived class.
-        """
-
-        if self.name_line is None:
-            raise EnvError("name_line is None. Backend is invalid. Line names are used to make the correspondance between the chronics and the backend")
-        if self.name_load is None:
-            raise EnvError("name_load is None. Backend is invalid. Line names are used to make the correspondance between the chronics and the backend")
-        if self.name_gen is None:
-            raise EnvError("name_gen is None. Backend is invalid. Line names are used to make the correspondance between the chronics and the backend")
-        if self.name_sub is None:
-            raise EnvError("name_sub is None. Backend is invalid. Line names are used to make the correspondance between the chronics and the backend")
-
-        # test if vector can be properly converted
-        if not isinstance(self.sub_info, np.ndarray):
-            try:
-                self.sub_info = np.array(self.sub_info)
-                self.sub_info = self.sub_info.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.sub_info should be convertible to a numpy array")
-
-        if not isinstance(self.load_to_subid, np.ndarray):
-            try:
-                self.load_to_subid = np.array(self.load_to_subid)
-                self.load_to_subid = self.load_to_subid.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.load_to_subid should be convertible to a numpy array")
-        if not isinstance(self.gen_to_subid, np.ndarray):
-            try:
-                self.gen_to_subid = np.array(self.gen_to_subid)
-                self.gen_to_subid = self.gen_to_subid.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.gen_to_subid should be convertible to a numpy array")
-        if not isinstance(self.line_or_to_subid, np.ndarray):
-            try:
-                self.line_or_to_subid = np.array(self.line_or_to_subid)
-                self.line_or_to_subid = self.line_or_to_subid .astype(np.int)
-            except Exception as e:
-                raise EnvError("self.line_or_to_subid should be convertible to a numpy array")
-        if not isinstance(self.line_ex_to_subid, np.ndarray):
-            try:
-                self.line_ex_to_subid = np.array(self.line_ex_to_subid)
-                self.line_ex_to_subid = self.line_ex_to_subid.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.line_ex_to_subid should be convertible to a numpy array")
-
-        if not isinstance(self.load_to_sub_pos, np.ndarray):
-            try:
-                self.load_to_sub_pos = np.array(self.load_to_sub_pos)
-                self.load_to_sub_pos = self.load_to_sub_pos.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.load_to_sub_pos should be convertible to a numpy array")
-        if not isinstance(self.gen_to_sub_pos, np.ndarray):
-            try:
-                self.gen_to_sub_pos = np.array(self.gen_to_sub_pos)
-                self.gen_to_sub_pos = self.gen_to_sub_pos.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.gen_to_sub_pos should be convertible to a numpy array")
-        if not isinstance(self.line_or_to_sub_pos, np.ndarray):
-            try:
-                self.line_or_to_sub_pos = np.array(self.line_or_to_sub_pos)
-                self.line_or_to_sub_pos = self.line_or_to_sub_pos.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.line_or_to_sub_pos should be convertible to a numpy array")
-        if not isinstance(self.line_ex_to_sub_pos, np.ndarray):
-            try:
-                self.line_ex_to_sub_pos = np.array(self.line_ex_to_sub_pos)
-                self.line_ex_to_sub_pos = self.line_ex_to_sub_pos .astype(np.int)
-            except Exception as e:
-                raise EnvError("self.line_ex_to_sub_pos should be convertible to a numpy array")
-
-        if not isinstance(self.load_pos_topo_vect, np.ndarray):
-            try:
-                self.load_pos_topo_vect = np.array(self.load_pos_topo_vect)
-                self.load_pos_topo_vect = self.load_pos_topo_vect.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.load_pos_topo_vect should be convertible to a numpy array")
-        if not isinstance(self.gen_pos_topo_vect, np.ndarray):
-            try:
-                self.gen_pos_topo_vect = np.array(self.gen_pos_topo_vect)
-                self.gen_pos_topo_vect = self.gen_pos_topo_vect.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.gen_pos_topo_vect should be convertible to a numpy array")
-        if not isinstance(self.line_or_pos_topo_vect, np.ndarray):
-            try:
-                self.line_or_pos_topo_vect = np.array(self.line_or_pos_topo_vect)
-                self.line_or_pos_topo_vect = self.line_or_pos_topo_vect.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.line_or_pos_topo_vect should be convertible to a numpy array")
-        if not isinstance(self.line_ex_pos_topo_vect, np.ndarray):
-            try:
-                self.line_ex_pos_topo_vect = np.array(self.line_ex_pos_topo_vect)
-                self.line_ex_pos_topo_vect = self.line_ex_pos_topo_vect.astype(np.int)
-            except Exception as e:
-                raise EnvError("self.line_ex_pos_topo_vect should be convertible to a numpy array")
-
-        # test that all numbers are finite:
-        tmp = np.concatenate((
-            self.sub_info.flatten(),
-                             self.load_to_subid.flatten(),
-                             self.gen_to_subid.flatten(),
-                             self.line_or_to_subid.flatten(),
-                             self.line_ex_to_subid.flatten(),
-                             self.load_to_sub_pos.flatten(),
-                             self.gen_to_sub_pos.flatten(),
-                             self.line_or_to_sub_pos.flatten(),
-                             self.line_ex_to_sub_pos.flatten(),
-                             self.load_pos_topo_vect.flatten(),
-                             self.gen_pos_topo_vect.flatten(),
-                             self.line_or_pos_topo_vect.flatten(),
-                             self.line_ex_pos_topo_vect.flatten()
-                              ))
-        try:
-            if np.any(~np.isfinite(tmp)):
-                raise EnvError("One of the vector is made of non finite elements")
-        except Exception as e:
-            raise EnvError("Impossible to check wheter or not vectors contains online finite elements (pobably one or more topology related vector is not valid (None)")
-
-        # check sizes
-        if len(self.sub_info) != self.n_sub:
-            raise IncorrectNumberOfSubstation("The number of substation is not consistent in self.sub_info (size \"{}\") and  self.n_sub ({})".format(len(self.sub_info), self.n_sub))
-        if np.sum(self.sub_info) != self.n_load + self.n_gen + 2*self.n_line:
-            err_msg = "The number of elements of elements is not consistent between self.sub_info where there are "
-            err_msg +=  "{} elements connected to all substations and the number of load, generators and lines in the _grid."
-            err_msg = err_msg.format(np.sum(self.sub_info))
-            raise IncorrectNumberOfElements(err_msg)
-
-        if len(self.load_to_subid) != self.n_load:
-            raise IncorrectNumberOfLoads()
-        if len(self.gen_to_subid) != self.n_gen:
-            raise IncorrectNumberOfGenerators()
-        if len(self.line_or_to_subid) != self.n_line:
-            raise IncorrectNumberOfLines()
-        if len(self.line_ex_to_subid) != self.n_line:
-            raise IncorrectNumberOfLines()
-
-        if len(self.load_to_sub_pos) != self.n_load:
-            raise IncorrectNumberOfLoads()
-        if len(self.gen_to_sub_pos) != self.n_gen:
-            raise IncorrectNumberOfGenerators()
-        if len(self.line_or_to_sub_pos) != self.n_line:
-            raise IncorrectNumberOfLines()
-        if len(self.line_ex_to_sub_pos) != self.n_line:
-            raise IncorrectNumberOfLines()
-
-        if len(self.load_pos_topo_vect) != self.n_load:
-            raise IncorrectNumberOfLoads()
-        if len(self.gen_pos_topo_vect) != self.n_gen:
-            raise IncorrectNumberOfGenerators()
-        if len(self.line_or_pos_topo_vect) != self.n_line:
-            raise IncorrectNumberOfLines()
-        if len(self.line_ex_pos_topo_vect) != self.n_line:
-            raise IncorrectNumberOfLines()
-
-        # test if object are connected to right substation
-        obj_per_sub = np.zeros(shape=(self.n_sub,))
-        for sub_id in self.load_to_subid:
-            obj_per_sub[sub_id] += 1
-        for sub_id in self.gen_to_subid:
-            obj_per_sub[sub_id] += 1
-        for sub_id in self.line_or_to_subid:
-            obj_per_sub[sub_id] += 1
-        for sub_id in self.line_ex_to_subid:
-            obj_per_sub[sub_id] += 1
-
-        if not np.all(obj_per_sub == self.sub_info):
-            raise IncorrectNumberOfElements()
-
-        # test right number of element in substations
-        # test that for each substation i don't have an id above the number of element of a substations
-        for i, (sub_id, sub_pos) in enumerate(zip(self.load_to_subid, self.load_to_sub_pos)):
-            if sub_pos >= self.sub_info[sub_id]:
-                raise IncorrectPositionOfLoads("for load {}".format(i))
-        for i, (sub_id, sub_pos) in enumerate(zip(self.gen_to_subid, self.gen_to_sub_pos)):
-            if sub_pos >= self.sub_info[sub_id]:
-                raise IncorrectPositionOfGenerators("for generator {}".format(i))
-        for i, (sub_id, sub_pos) in enumerate(zip(self.line_or_to_subid, self.line_or_to_sub_pos)):
-            if sub_pos >= self.sub_info[sub_id]:
-                raise IncorrectPositionOfLines("for line {} at origin end".format(i))
-        for i, (sub_id, sub_pos) in enumerate(zip(self.line_ex_to_subid, self.line_ex_to_sub_pos)):
-            if sub_pos >= self.sub_info[sub_id]:
-                # pdb.set_trace()
-                raise IncorrectPositionOfLines("for line {} at extremity end".format(i))
-
-        # check that i don't have 2 objects with the same id in the "big topo" vector
-        if len(np.unique(np.concatenate((self.load_pos_topo_vect.flatten(),
-                                        self.gen_pos_topo_vect.flatten(),
-                                        self.line_or_pos_topo_vect.flatten(),
-                                        self.line_ex_pos_topo_vect.flatten())))) != np.sum(self.sub_info):
-                raise EnvError("2 different objects would have the same id in the topology vector.")
-
-        # check that self.load_pos_topo_vect and co are consistent
-        load_pos_big_topo = self._aux_pos_big_topo(self.load_to_subid, self.load_to_sub_pos)
-        if not np.all(load_pos_big_topo == self.load_pos_topo_vect):
-            raise IncorrectPositionOfLoads()
-        gen_pos_big_topo = self._aux_pos_big_topo(self.gen_to_subid, self.gen_to_sub_pos)
-        if not np.all(gen_pos_big_topo == self.gen_pos_topo_vect):
-            raise IncorrectNumberOfGenerators()
-        lines_or_pos_big_topo = self._aux_pos_big_topo(self.line_or_to_subid, self.line_or_to_sub_pos)
-        if not np.all(lines_or_pos_big_topo == self.line_or_pos_topo_vect):
-            raise IncorrectPositionOfLines()
-        lines_ex_pos_big_topo = self._aux_pos_big_topo(self.line_ex_to_subid, self.line_ex_to_sub_pos)
-        if not np.all(lines_ex_pos_big_topo == self.line_ex_pos_topo_vect):
-            raise IncorrectPositionOfLines()
-
-        # no empty bus: at least one element should be present on each bus
-        if np.any(self.sub_info < 1):
-            raise BackendError("There are {} bus with 0 element connected to it.".format(np.sum(self.sub_info < 1)))
 
     def assert_grid_correct_after_powerflow(self):
         """
